--- src/play/lib/client/index.ts
@@ -1,1 +1,0 @@
-export * from './client';
