--- src/play/lib/server/index.ts
@@ -1,2 +1,0 @@
-export { initTRPC } from './initTRPC';
-export { ProceduresByType } from './router';
