+++ src/trpc/server/index.ts
@@ -1,5 +1,2 @@
-export * from './procedure';
-export * from './initTRPC';
-export * from './middlewares';
-export * from './router';
-export * from './types';
+export { initTRPC } from './initTRPC';
+export { ProceduresByType } from './router';
