--- src/trpc/server/middlewares/index.ts
@@ -1,2 +1,0 @@
-export * from './newContext';
-export * from './zod';
